requirements.txt
joblib==1.4.2
pandas==2.3.3
Pillow==12.0.0
scikit_learn==1.7.2
selenium==4.38.0
torch==2.9.0
transformers==4.50.3
webdriver_manager==4.0.2
xgboost==3.1.1
